requirements.txt
anyio==4.2.0
astor==0.8.1
async-timeout==4.0.3
asyncpg==0.29.0
black==23.11.0
click==8.1.6
exceptiongroup==1.2.0
fastapi==0.95.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pydantic==1.10.14
pynest-api==0.1.2
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.23.1
